pyproject.toml
[project]
name = "office-jinja"
version = "0.1.0"
description = ""
authors = [
    {name = "tolatolatop",email = "tolatolatop@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0.0"
dependencies = [
    "docxtpl (>=0.20.2,<0.21.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pydantic (>=2.12.4,<3.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "click (>=8.3.1,<9.0.0)",
    "rich (>=14.2.0,<15.0.0)",
    "xlsxwriter (>=3.2.9,<4.0.0)",
    "pydantic-yaml (>=1.6.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "office_jinja", from = "src"}]


[tool.poetry.group.dev.dependencies]
black = "^25.11.0"
isort = "^7.0.0"
ruff = "^0.14.6"
mypy = "^1.18.2"
pytest = "^9.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
office-jinja = "office_jinja.cli:generate"
